#565 [feature] Change Comment Dto

+++ data/src/main/java/daily/dayo/data/datasource/remote/comment/CommentApiService.kt
@@ -20,5 +20,5 @@ interface CommentApiService {
 
     /*** v1 ***/
     @POST("/api/v1/comments/delete/{commentId}")
-    suspend fun requestDeletePostComment(@Path("commentId") commentId: Int): NetworkResponse<Void>
+    suspend fun requestDeletePostComment(@Path("commentId") commentId: Long): NetworkResponse<Void>
 }
